isbn_mo/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.0.27
python-dotenv
